requirements-dev.txt
pytest>=7.0
pytest-cov>=4.1
pyyaml>=6.0

